Add material indicator delegate to readme
